requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.7.0
multidict==6.0.4
openai==0.28.1
PyYAML==6.0.1
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==2.0.5
yarl==1.9.2
zipp==3.15.0
